+++ code/examples/actor/parts/src/lib.rs
@@ -1,8 +1,9 @@
 pub mod actor;
 pub mod codec;
 pub mod config;
+pub mod consensus_host;
 pub mod db_metrics;
-pub mod mempool;
+pub mod fifo_mempool;
 pub mod mempool_load;
 pub mod metrics;
 pub mod mock_host;