+++ code/examples/actor/parts/src/types/signing.rs
@@ -1,3 +1,5 @@
+use async_trait::async_trait;
+
 use bytes::Bytes;
 
 use malachitebft_core_types::{
@@ -49,13 +51,14 @@ impl Ed25519Provider {
     }
 }
 
+#[async_trait]
 impl SigningProvider<MockContext> for Ed25519Provider {
-    fn sign_vote(&self, vote: Vote) -> SignedVote<MockContext> {
+    async fn sign_vote(&self, vote: Vote) -> SignedVote<MockContext> {
         let signature = self.sign(&vote.to_sign_bytes());
         SignedVote::new(vote, signature)
     }
 
-    fn verify_signed_vote(
+    async fn verify_signed_vote(
         &self,
         vote: &Vote,
         signature: &Signature,
@@ -64,12 +67,12 @@ impl SigningProvider<MockContext> for Ed25519Provider {
         public_key.verify(&vote.to_sign_bytes(), signature).is_ok()
     }
 
-    fn sign_proposal(&self, proposal: Proposal) -> SignedProposal<MockContext> {
+    async fn sign_proposal(&self, proposal: Proposal) -> SignedProposal<MockContext> {
         let signature = self.private_key.sign(&proposal.to_sign_bytes());
         SignedProposal::new(proposal, signature)
     }
 
-    fn verify_signed_proposal(
+    async fn verify_signed_proposal(
         &self,
         proposal: &Proposal,
         signature: &Signature,
@@ -80,12 +83,15 @@ impl SigningProvider<MockContext> for Ed25519Provider {
             .is_ok()
     }
 
-    fn sign_proposal_part(&self, proposal_part: ProposalPart) -> SignedProposalPart<MockContext> {
+    async fn sign_proposal_part(
+        &self,
+        proposal_part: ProposalPart,
+    ) -> SignedProposalPart<MockContext> {
         let signature = self.private_key.sign(&proposal_part.to_sign_bytes());
         SignedProposalPart::new(proposal_part, signature)
     }
 
-    fn verify_signed_proposal_part(
+    async fn verify_signed_proposal_part(
         &self,
         proposal_part: &ProposalPart,
         signature: &Signature,
@@ -96,12 +102,12 @@ impl SigningProvider<MockContext> for Ed25519Provider {
             .is_ok()
     }
 
-    fn sign_vote_extension(&self, extension: Bytes) -> SignedExtension<MockContext> {
+    async fn sign_vote_extension(&self, extension: Bytes) -> SignedExtension<MockContext> {
         let signature = self.private_key.sign(extension.as_ref());
         malachitebft_core_types::SignedMessage::new(extension, signature)
     }
 
-    fn verify_signed_vote_extension(
+    async fn verify_signed_vote_extension(
         &self,
         extension: &Bytes,
         signature: &Signature,