+++ RELEASE_NOTES.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
-> Nothing yet
+- Make SigningProvider trait methods async ([#1151](https://github.com/informalsystems/malachite/issues/1151))
+- Make GossipSub topic names configurable ([#849](https://github.com/informalsystems/malachite/issues/849))
+- Fix bug in WAL recovery logic where a corrupted entry would not be detected in some circumstances ([#1127](https://github.com/informalsystems/malachite/pull/1127))
 
 ## 0.5.0
 
@@ -13,6 +15,7 @@
 - Decouple `Host` messages from the `Consensus` actor ([#1109](https://github.com/informalsystems/malachite/pull/1109))
 - Fix a bug where values synced from other peers were assigned the current node's address instead of their proposer's address ([#1141](https://github.com/informalsystems/malachite/pull/1141))
 - Buffer sync values for heights higher than current height in consensus and replay when running consensus for those heights ([#1149](https://github.com/informalsystems/malachite/pull/1149))
+- Add value batching to sync messages ([#1070](https://github.com/informalsystems/malachite/issues/1070))
 
 ## 0.4.0